requirements.txt
streamlit==1.25.0
ultralytics==8.3.59
paddleocr==2.9.1
paddlepaddle==2.6.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
matplotlib==3.7.3
Pillow==9.0.1
scikit-learn==1.3.2
scikit-image==0.22.0
scipy==1.11.4
protobuf==3.20.3

torch==2.3.1
transformers==4.49.0
datasets==3.4.1
sentencepiece==0.2.0

IPython==8.12.0

llama-cpp-python==0.3.8